+++ src/com/geoscope/GeoEye/Space/TypesSystem/TileServerVisualization/TTileServerVisualizationFunctionality.java
@@ -12,7 +12,7 @@ import com.geoscope.GeoEye.Space.TypesSystem.TComponentData;
 
 public class TTileServerVisualizationFunctionality extends TComponentFunctionality {
 
-	public TTileServerVisualizationFunctionality(TTypeFunctionality pTypeFunctionality, int pidComponent) {
+	public TTileServerVisualizationFunctionality(TTypeFunctionality pTypeFunctionality, long pidComponent) {
 		super(pTypeFunctionality,pidComponent);
 	}
 	
@@ -21,7 +21,7 @@ public class TTileServerVisualizationFunctionality extends TComponentFunctionali
 		String URL1 = Server.Address;
 		//. add command path
 		URL1 = "http://" + URL1 + "/" + "Space" + "/" + "2"/* URLProtocolVersion */ + "/" + Integer.toString(Server.User.UserID);
-		String URL2 = "TypesSystem" + "/" + Integer.toString(TypeFunctionality.idType) + "/" + "Co" + "/" + Integer.toString(idComponent) + "/" + "Data.dat";
+		String URL2 = "TypesSystem" + "/" + Integer.toString(TypeFunctionality.idType) + "/" + "Co" + "/" + Long.toString(idComponent) + "/" + "Data.dat";
 		//. add command parameters
 		URL2 = URL2 + "?" + "0"/* command */+ "," + "1"/* version of data*/;
 		//.
